improve sdk-remote-proxy setup
