+++ examples/remote-proxy/.papi/descriptors/.gitignore
