+++ examples/remote-proxy/.papi/descriptors/package.json
