+++ examples/remote-proxy/.papi/metadata/ksm.scale
