+++ examples/remote-proxy/.papi/metadata/ksmAh.scale
