+++ examples/remote-proxy/.papi/metadata/wnd-metadata.scale
