+++ examples/remote-proxy/index.html
@@ -1,0 +1,13 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>polkadot-api-legacy-provider</title>
+  </head>
+
+  <body style="display: none">
+    <script type="module" src="/src/sign-tx.ts"></script>
+  </body>
+</html>
