+++ examples/remote-proxy/package.json
