+++ examples/remote-proxy/src/vite-env.d.ts
