+++ examples/remote-proxy/vite.config.ts
