+++ packages/sdk-remote-proxy/.papi/descriptors/package.json
@@ -1,5 +1,5 @@
 {
-  "version": "0.1.0-autogenerated.12310711838707458573",
+  "version": "0.1.0-autogenerated.11708617951874670738",
   "name": "@polkadot-api/descriptors",
   "files": [
     "dist"
