+++ packages/sdk-remote-proxy/.prettierignore
@@ -1,0 +1,1 @@
+src/descriptors
